requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.3.0
azure-cognitiveservices-speech==1.37.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.7.0
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
mysqlclient==2.2.4
ormar==0.20.0
pydantic==2.5.3
pydantic-core==2.14.6
pymysql==1.1.0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
sniffio==1.3.1
sqlalchemy==1.4.52
starlette==0.37.2
typing-extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.3
pyjwt==2.8.0
pytest==8.1.1
python-jose==3.3.0